♻️ Move cli logic to src/cli.js
